Todos x id de usuario y fin de clases

+++ src/app/api/todos/route.ts
@@ -1,4 +1,5 @@
 import prisma from '@/app/lib/prisma'
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import { NextResponse, NextRequest } from 'next/server'
 import * as yup from 'yup';
 
@@ -33,6 +34,12 @@ const postSchema = yup.object({
 
 export async function POST(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) {
+        return NextResponse.json('No autorizado', { status: 401 });
+    }
+
     try {
 
         // const body = await postSchema.validate(await request.json()); // Forma 1 
@@ -42,27 +49,33 @@ export async function POST(request: Request) {
          * todo: const { complete, description, ...otrasPropiedades }
          * todo: De esta forma puedo darle manejo u errores si salen otras propiedades en el body
          */
-        
-        const { complete, description } = await postSchema.validate(await request.json()); 
+
+        const { complete, description } = await postSchema.validate(await request.json());
 
         // const todo = await prisma.todo.create({ data: body }); // Forma 1
-        const todo = await prisma.todo.create({ data: {complete, description} });
+        const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } });
 
         return NextResponse.json(todo);
 
     } catch (error) {
-        return NextResponse.json(error, {status: 400});
+        return NextResponse.json(error, { status: 400 });
     }
 }
 
 export async function DELETE(request: Request) {
 
+    const user = await getUserSessionServer();
+
+    if (!user) {
+        return NextResponse.json('No autorizado', { status: 401 });
+    }
+
     try {
-        const todo = await prisma.todo.deleteMany({where: {complete: true}});
+        const todo = await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
 
         return NextResponse.json('Borrados');
 
     } catch (error) {
-        return NextResponse.json(error, {status: 400});
+        return NextResponse.json(error, { status: 400 });
     }
 }
